book_recommender_api/requirements.txt
# requirements.txt
fastapi==0.95.2
uvicorn==0.29.0
pymongo==4.6.1
pydantic==1.10.13
python-dotenv==1.0.1